requirements.txt
google-cloud-texttospeech
google-cloud-translate
pysrt
